+++ src/Engine2D/Rendering/Camera2D.cpp
@@ -18,21 +18,7 @@ namespace Engine2D::Rendering {
   Camera2D::Camera2D(
     const float left, const float right, const float bottom, const float top, const float near, const float far
   )
-    : positionOffset(0), rotationOffset(0), damping(0), shakeCoefficients(
-        {
-          std::vector<ShakeWave>{
-            {5.5f, 20.8f, 2.3f},
-            {2.2f, 5.6f, 5.0f},
-            {1.8f, 11.2f, 1.0f}
-          },
-          std::vector<ShakeWave>{
-            {4.0f, 17.6f, -2.5f},
-            {2.8f, 9.6f, 4.0f},
-            {1.6f, 1.6f, 0.0f}
-          }
-        }
-      ),
-      projection(glm::ortho(left, right, bottom, top, near, far)),
+    : positionOffset(0), rotationOffset(0), damping(0), projection(glm::ortho(left, right, bottom, top, near, far)),
       view(1.0f), shaking(false), shakeElapsed(0), shakeDuration(0), m00(0), m01(0), m03(0), m10(0), m11(0), m13(0) {
     viewProjection = projection * view;
   }
@@ -130,10 +116,10 @@ namespace Engine2D::Rendering {
 
     float x = 0.0f, y = 0.0f;
 
-    for (const auto &[amplitude, frequency, phase]: shakeCoefficients[0])
+    for (const auto &[amplitude, frequency, phase]: shakeCoefficientsX)
       x += amplitude * std::sin(frequency * glm::pi<float>() * frac + phase);
 
-    for (const auto &[amplitude, frequency, phase]: shakeCoefficients[1])
+    for (const auto &[amplitude, frequency, phase]: shakeCoefficientsY)
       y += amplitude * std::sin(frequency * glm::pi<float>() * frac + phase);
 
     const float lim = (frac <= mid) ? frac / mid : 1.0f - (frac - mid) / (1.0f - mid);